+++ src/index.ts
@@ -3,6 +3,7 @@ import './components/component-island';
 import './components/fragment-island';
 import './components/layout-island';
 import './components/page-island';
+import { setupNavigation } from './navigation';
 
 type CastawayAppOptions = {
     components?: ComponentDefinition
@@ -10,6 +11,7 @@ type CastawayAppOptions = {
 
 export const createCastawayApp = (options: CastawayAppOptions) => {
     defineComponents(options.components ?? {});
+    setupNavigation();
 };
 
 export default createCastawayApp;